Avoid NullPointerException if File.listFiles() returns null !

+++ dcm4chee/dcm4chee-arc/branches/B_WITHOUT_RID/dcm4jboss-wado/src/java/org/dcm4chex/wado/mbean/cache/WADOCacheImpl.java
@@ -253,6 +253,7 @@ public class WADOCacheImpl implements WADOCache {
 		if ( !file.isDirectory() ) return;
 		
 		File[] files = file.listFiles();
+		if ( files == null ) return;
 		for ( int i = 0, len = files.length ; i < len ; i++ ) {
 			if ( files[i].isFile() ) {
 				files[i].delete();
